% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtaware.R
\name{get_first_column_of_dataframe}
\alias{get_first_column_of_dataframe}
\title{Get the first column of a data.frame}
\usage{
get_first_column_of_dataframe(df)
}
\arguments{
\item{df}{a data.frame or data.table}
}
\value{
\code{df[1]}, the first column, as a data.frame/table
}
\description{
Gets the first column of a data.frame OR data.table,
as a data.frame/table, by running \code{df[1]} with
.datatable.aware = FALSE
}
